Move the factories to spec/support/ folder
