+++ spec/factory_spec.rb
@@ -53,18 +53,4 @@ describe 'Factory' do
       })
     end
   end
-
-  def factories
-    Falcor::Fabricator.define :author do
-      field :username, "matt@example.com"
-      field :password, "password"
-    end
-
-    Falcor::Fabricator.define :blog_post do
-      field :title, "My post"
-      field :body, "Lots of text"
-      association :author
-    end
-  end
-
 end
